Commit: Setup and added initial structure
Explain:
I have figured out the structure of MySPlit. All of the forms and models will be located in the general app and all of the development testing views will be added to the geneal app as well. Each of the apps will just contain the views and the urls for that specific app. Templates will also be stored in the general application of MySplit